+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.15.0] 2022-09-27
 
+### Added
+- New editor for `langlabel` properties.
+
 ### Fixed
 - Typo in settings panel.
 - Fixed compatibility with the newer Electron versions.